DB - course table schema update, using uuid as primary key
