+++ package.json
@@ -27,7 +27,8 @@
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.52.0",
-    "react-icons": "^4.12.0"
+    "react-icons": "^4.12.0",
+    "uuid": "^10.0.0"
   },
   "devDependencies": {
     "@types/dompurify": "^3.0.5",
