+++ src/app/utilities/db/course/describeCourseTable.ts
@@ -1,6 +1,6 @@
 import { DescribeTableCommand } from '@aws-sdk/client-dynamodb';
-import { dbDocumentClient } from './configure';
-import { COURSE_TABLE_NAME } from './constants';
+import { dbDocumentClient } from '../configure';
+import { COURSE_TABLE_NAME } from '../constants';
 
 export default async function describeCourseTable() {
   return await dbDocumentClient.send(
