+++ src/app/utilities/db/course/insertCourseRows.ts
@@ -1,6 +1,6 @@
 import { BatchWriteCommand } from '@aws-sdk/lib-dynamodb';
-import { COURSE_TABLE_NAME } from './constants';
-import { dbDocumentClient } from './configure';
+import { COURSE_TABLE_NAME } from '../constants';
+import { dbDocumentClient } from '../configure';
 import { CourseProps } from '@/app/interfaces/Course';
 
 export default async function insertCourseRows(courses: CourseProps[]) {
